set type int64 -> int32

+++ pkg/structure/id.go
@@ -6,10 +6,10 @@ import (
 	"strings"
 )
 
-func GenSnapshotID(targetChainID string, targetChainHeight int64, initTimestamp Timestamp) string {
+func GenSnapshotID(targetChainID string, targetChainHeight int32, initTimestamp Timestamp) string {
 	return strings.Join([]string{
 		targetChainID,
-		strconv.FormatInt(targetChainHeight, 10),
+		strconv.FormatInt(int64(targetChainHeight), 10),
 		initTimestamp.Format(constants.TIME_LAYOUT_DAY)}, constants.ID_SEP_SYMBOL)
 }
 
@@ -19,8 +19,8 @@ func GenEventID(chainID string, eventTag string) string {
 
 // GenChainID 根据当前链所处的 fork 的区块高度作为该 chain 的唯一标识，
 // TODO 是否存在不同 fork 高度相同的情况
-func GenChainID(forkHeight int64) string {
-	return strconv.FormatInt(forkHeight, 10)
+func GenChainID(forkHeight int32) string {
+	return strconv.FormatInt(int64(forkHeight), 10)
 }
 
 func ParseChainIDFromSnapshotID(snapshotID string) string {
